requirements.txt
langchain-google-genai
chainlit
python-dotenv
langchain